% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{flame_graph}
\alias{flame_graph}
\title{Flame graph}
\usage{
flame_graph(trace_data)
}
\arguments{
\item{trace_data}{a data.frame, output of run_trace()}
}
\value{
a ggplot object
}
\description{
Create a flame graph of trace data
}
\examples{
\dontrun{
df <- run_trace(ggplot2::ggplot(), c("ggplot2", "scales"))
flame_graph(df)
}
}
