% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{recolour}
\alias{recolour}
\title{Conditionally add crayon colours to a text vector}
\usage{
recolour(text, group, ..., .default = "black")
}
\arguments{
\item{text}{a vector of text}

\item{group}{a vector of values to}

\item{...}{a list of pairs: \if{html}{\out{<value in group>}} = \if{html}{\out{<colour>}}}

\item{.default}{the default colour for groups not specified}
}
\value{
a character vector same length as text
}
\description{
Think of this as dplyr::recode, but for adding colours. Colours are added
based on the group variable
}
\examples{
cat(recolour(c("hello ", "world!"), c("a", "b"), "a" = "blue", "b" = "red"))
}
