% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{run_trace}
\alias{run_trace}
\title{Trace how an expression uses packages}
\usage{
run_trace(expr, packages)
}
\arguments{
\item{expr}{an expression}

\item{packages}{a character vector of packages (cannot include base package)}
}
\value{
a data.frame
}
\description{
This evaluates an expression, and traces where functions from specified
packages are used.
}
\examples{
\dontrun{
df <- run_trace(dplyr::if_else(TRUE, 1, 100), "rlang")
}
}
