% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{trace_package}
\alias{trace_package}
\alias{untrace_package}
\title{Trace functions in a package}
\usage{
trace_package(pkg_name)

untrace_package(pkg_name)
}
\arguments{
\item{pkg_name}{single package name}
}
\value{
the functions traced (invisibly)
}
\description{
This adds (and removes) traces to the functions of a package: whenever those
functions are called, some diagnostic information is printed to the console.
}
\details{
Packages must be loaded and attached to be traced: this function ensures
this and provides warnings about packages that can't be loaded.
}
\note{
This trace does not work with the base package
}
\examples{
trace_package("ggplot2")
}
