% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{code_tree}
\alias{code_tree}
\title{Print a tree of traced function calls}
\usage{
code_tree(trace_data, label = call)
}
\arguments{
\item{trace_data}{a data.frame, output of run_trace()}

\item{label}{the label used at each node. By default, the \code{call} for each
function call traced.}
}
\value{
invisible(NULL)
}
\description{
Print a tree of traced function calls
}
\examples{
df <- run_trace(ggplot2::ggplot(), c("ggplot2", "scales"))
code_tree(df)

code_tree(df, fun)

code_tree(df, recolour(fun, pkg, ggplot2 = "red", scales = "blue"))
}
